@@ -1,33 +1,29 @@
 {
   "name": "parenting-app-ui",
-  "version": "0.5.1",
-  "author": "Ionic Framework",
-  "homepage": "https://ionicframework.com/",
+  "version": "0.5.3",
+  "author": "IDEMS International",
+  "homepage": "https://idems.international/",
+  "description": "",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --open",
     "build": "ng build --prod",
-    "build-netlify": "npm run populate-netlify-secrets && npm run build",
+    "scripts": "cd scripts && npm run start --",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "version": "ts-node --project tsconfig.scripts.json scripts/version.ts",
-    "icon": "ts-node --project tsconfig.scripts.json scripts/icon.ts",
-    "populate-netlify-secrets": "echo $FIREBASE_CONFIG_TS > src/environments/firebaseConfig.prod.ts && echo $FIREBASE_CONFIG_TS > src/environments/firebaseConfig.ts",
-    "process-content-spreadsheet": "ts-node --project tsconfig.scripts.json scripts/plh-spreadsheet/process-content-spreadsheet.ts",
-    "xlsx-to-json": "ts-node --project tsconfig.scripts.json scripts/plh-xlsx-to-json/index.ts",
-    "test-scripts": "jasmine-ts --project tsconfig.scripts.json --config=scripts.jasmine.json"
+    "postinstall": "ngcc && cd scripts && yarn install"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.2.3",
-    "@angular/common": "~10.1.4",
-    "@angular/core": "~10.1.4",
+    "@angular/animations": "^11.0.2",
+    "@angular/common": "~11.0.2",
+    "@angular/core": "~11.0.2",
     "@angular/fire": "^5.4.2",
-    "@angular/forms": "~10.1.4",
-    "@angular/platform-browser": "~10.1.4",
-    "@angular/platform-browser-dynamic": "~10.1.4",
-    "@angular/router": "~10.1.4",
+    "@angular/forms": "~11.0.2",
+    "@angular/platform-browser": "~11.0.2",
+    "@angular/platform-browser-dynamic": "~11.0.2",
+    "@angular/router": "~11.0.2",
     "@capacitor/android": "^2.4.1",
     "@capacitor/cli": "^2.4.1",
     "@capacitor/core": "^2.4.1",
