@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { AppTermsPage } from './app-terms.page';
@@ -7,7 +7,7 @@ describe('AppTermsPage', () => {
   let component: AppTermsPage;
   let fixture: ComponentFixture<AppTermsPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AppTermsPage ],
       imports: [IonicModule.forRoot()]
