@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { ChatResponsesComponent } from './chat-responses.component';
@@ -7,7 +7,7 @@ describe('ChatResponsesComponent', () => {
   let component: ChatResponsesComponent;
   let fixture: ComponentFixture<ChatResponsesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ChatResponsesComponent ],
       imports: [IonicModule.forRoot()]
