@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { ChatPage } from './chat.page';
@@ -7,7 +7,7 @@ describe('ChatPage', () => {
   let component: ChatPage;
   let fixture: ComponentFixture<ChatPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ChatPage ],
       imports: [IonicModule.forRoot()]
