@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { AddFamilyMemberPage } from './add-family-member.page';
@@ -7,7 +7,7 @@ describe('AddFamilyMemberPage', () => {
   let component: AddFamilyMemberPage;
   let fixture: ComponentFixture<AddFamilyMemberPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AddFamilyMemberPage ],
       imports: [IonicModule.forRoot()]
