@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { FamilyPage } from './family.page';
@@ -7,7 +7,7 @@ describe('FamilyPage', () => {
   let component: FamilyPage;
   let fixture: ComponentFixture<FamilyPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FamilyPage ],
       imports: [IonicModule.forRoot()]
