@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { GalleryPage } from './gallery.page';
@@ -7,7 +7,7 @@ describe('GalleryPage', () => {
   let component: GalleryPage;
   let fixture: ComponentFixture<GalleryPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ GalleryPage ],
       imports: [IonicModule.forRoot()]
