@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { IonicModule } from "@ionic/angular";
 
 import { EditReminderComponent } from "./edit-reminder.component";
@@ -7,7 +7,7 @@ describe("EditReminderComponent", () => {
   let component: EditReminderComponent;
   let fixture: ComponentFixture<EditReminderComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [EditReminderComponent],
       imports: [IonicModule.forRoot()],
