@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { GoalsPage } from './goals.page';
@@ -7,7 +7,7 @@ describe('GoalsPage', () => {
   let component: GoalsPage;
   let fixture: ComponentFixture<GoalsPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ GoalsPage ],
       imports: [IonicModule.forRoot()]
