@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { HomeButtonsComponent } from './home-buttons.component';
@@ -7,7 +7,7 @@ describe('HomeButtonsComponent', () => {
   let component: HomeButtonsComponent;
   let fixture: ComponentFixture<HomeButtonsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ HomeButtonsComponent ],
       imports: [IonicModule.forRoot()]
