@@ -1,4 +1,5 @@
 export type UserSettingId = "CHAT_DELAY" | "USE_MODAL_FOR_CHAT_RESPONSES" | "USE_BUTTON_HOME_SCREEN" | "USE_OFFLINE_CHAT"
+    | "USE_GDRIVE_CONTENT"
     | "SHOW_FLOW_NAME";
 
 export interface UserSetting {
@@ -7,7 +8,9 @@ export interface UserSetting {
     type: "boolean" | "string" | "number",
     value: string,
     options?: UserSettingOption[]
-    nativeOnly?: boolean
+    nativeOnly?: boolean,
+    devOnly?: boolean,
+    /* Route to go to when button clicked */
 }
 
 export interface UserSettingOption {
@@ -40,19 +43,22 @@ export const BASE_USER_SETTINGS: UserSetting[] = [
         id: "USE_MODAL_FOR_CHAT_RESPONSES",
         name: "Use Modal For Chat Responses",
         type: "boolean",
-        value: "false"
+        value: "false",
+        devOnly: true,
     },
     {
-        id: "USE_BUTTON_HOME_SCREEN",
-        name: "Use Button Home Screen",
+        id: "USE_GDRIVE_CONTENT",
+        name: "Use Latest GDrive Content",
         type: "boolean",
-        value: "true"
+        value: "false",
+        devOnly: true
     },
     {
         id: "SHOW_FLOW_NAME",
         name: "Show Flow Name",
         type: "boolean",
-        value: "true"
+        value: "true",
+        devOnly: true
     }
     // 2020-11-25 - Online chat disabled here and in settings until tested working
     // {
