@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { BreathingAnimComponent } from './breathing-anim.component';
@@ -7,7 +7,7 @@ describe('BreathingAnimComponent', () => {
   let component: BreathingAnimComponent;
   let fixture: ComponentFixture<BreathingAnimComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BreathingAnimComponent ],
       imports: [IonicModule.forRoot()]
