@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { ListeningAnimComponent } from './listening-anim.component';
@@ -7,7 +7,7 @@ describe('ListeningAnimComponent', () => {
   let component: ListeningAnimComponent;
   let fixture: ComponentFixture<ListeningAnimComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ListeningAnimComponent ],
       imports: [IonicModule.forRoot()]
