@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseBodyScannerComponent } from './pause-body-scanner.component';
@@ -7,7 +7,7 @@ describe('PauseBodyScannerComponent', () => {
   let component: PauseBodyScannerComponent;
   let fixture: ComponentFixture<PauseBodyScannerComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseBodyScannerComponent ],
       imports: [IonicModule.forRoot()]
