@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseFeelComponent } from './pause-feel.component';
@@ -7,7 +7,7 @@ describe('PauseFeelComponent', () => {
   let component: PauseFeelComponent;
   let fixture: ComponentFixture<PauseFeelComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseFeelComponent ],
       imports: [IonicModule.forRoot()]
