@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseStep0Component } from './pause-step0.component';
@@ -7,7 +7,7 @@ describe('PauseStep0Component', () => {
   let component: PauseStep0Component;
   let fixture: ComponentFixture<PauseStep0Component>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseStep0Component ],
       imports: [IonicModule.forRoot()]
