@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseStep2Component } from './pause-step2.component';
@@ -7,7 +7,7 @@ describe('PauseStep2Component', () => {
   let component: PauseStep2Component;
   let fixture: ComponentFixture<PauseStep2Component>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseStep2Component ],
       imports: [IonicModule.forRoot()]
