@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseStep3Component } from './pause-step3.component';
@@ -7,7 +7,7 @@ describe('PauseStep3Component', () => {
   let component: PauseStep3Component;
   let fixture: ComponentFixture<PauseStep3Component>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseStep3Component ],
       imports: [IonicModule.forRoot()]
