@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseStep4Component } from './pause-step4.component';
@@ -7,7 +7,7 @@ describe('PauseStep4Component', () => {
   let component: PauseStep4Component;
   let fixture: ComponentFixture<PauseStep4Component>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseStep4Component ],
       imports: [IonicModule.forRoot()]
