@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseStep5Component } from './pause-step5.component';
@@ -7,7 +7,7 @@ describe('PauseStep5Component', () => {
   let component: PauseStep5Component;
   let fixture: ComponentFixture<PauseStep5Component>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseStep5Component ],
       imports: [IonicModule.forRoot()]
