@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { PauseThinkComponent } from './pause-think.component';
@@ -7,7 +7,7 @@ describe('PauseThinkComponent', () => {
   let component: PauseThinkComponent;
   let fixture: ComponentFixture<PauseThinkComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PauseThinkComponent ],
       imports: [IonicModule.forRoot()]
