@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { StressedAnimComponent } from './stressed-anim.component';
@@ -7,7 +7,7 @@ describe('StressedAnimComponent', () => {
   let component: StressedAnimComponent;
   let fixture: ComponentFixture<StressedAnimComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ StressedAnimComponent ],
       imports: [IonicModule.forRoot()]
