@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { ToolboxFlowsPage } from './toolbox-flows.page';
@@ -7,7 +7,7 @@ describe('ToolboxFlowsPage', () => {
   let component: ToolboxFlowsPage;
   let fixture: ComponentFixture<ToolboxFlowsPage>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ToolboxFlowsPage ],
       imports: [IonicModule.forRoot()]
