@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { BlobComponent } from './blob.component';
@@ -7,7 +7,7 @@ describe('BlobComponent', () => {
   let component: BlobComponent;
   let fixture: ComponentFixture<BlobComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BlobComponent ],
       imports: [IonicModule.forRoot()]
