@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { ReflectAnimationComponent } from './reflect-animation.component';
@@ -7,7 +7,7 @@ describe('ReflectAnimationComponent', () => {
   let component: ReflectAnimationComponent;
   let fixture: ComponentFixture<ReflectAnimationComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ReflectAnimationComponent ],
       imports: [IonicModule.forRoot()]
