@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
 
 import { StressedMultiHandAnimComponent } from './stressed-multi-hand-anim.component';
@@ -7,7 +7,7 @@ describe('StressedMultiHandAnimComponent', () => {
   let component: StressedMultiHandAnimComponent;
   let fixture: ComponentFixture<StressedMultiHandAnimComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ StressedMultiHandAnimComponent ],
       imports: [IonicModule.forRoot()]
