@@ -11,7 +11,7 @@ export function generateRandomId() {
  * Convert an object array into a json object, with keys corresponding to array entries
  * @param keyfield any unique field which all array objects contain to use as hash keys (e.g. 'id')
  */
-export function arrayToHashmap<T>(arr: T[], keyfield: string) {
+export function arrayToHashmap<T>(arr: T[], keyfield: string): { [key: string]: T } {
   const hashmap: { [key: string]: T } = {};
   for (const el of arr) {
     if (el.hasOwnProperty(keyfield)) {
